246b1757984c6e5df24743ceea25352770fb7c19

+++ storage/ndb/test/include/NdbProcess.hpp
@@ -1,5 +1,5 @@
 /*
-  Copyright (c) 2009, 2022, Oracle and/or its affiliates.
+  Copyright (c) 2009, 2023, Oracle and/or its affiliates.
 
 
    This program is free software; you can redistribute it and/or modify
@@ -137,9 +137,15 @@ public:
   {
 #ifdef _WIN32
     HANDLE processHandle = OpenProcess(PROCESS_ALL_ACCESS, FALSE, m_pid);
-    if(!TerminateProcess(processHandle,9999))
+    if (!processHandle)
     {
       printerror();
+      return false;
+    }
+    if (!TerminateProcess(processHandle,9999))
+    {
+      printerror();
+      CloseHandle(processHandle);
       return false;
     }
     CloseHandle(processHandle);
@@ -162,6 +168,11 @@ public:
   {
 #ifdef _WIN32
     HANDLE processHandle = OpenProcess(PROCESS_ALL_ACCESS, FALSE, m_pid);
+    if (!processHandle)
+    {
+      printerror();
+      return false;
+    }
     const DWORD result = WaitForSingleObject(processHandle, timeout*100);
     bool fun_ret = true;
     if (result == WAIT_TIMEOUT) {
@@ -178,8 +189,10 @@ public:
     {
       fprintf(stderr,
         "Error occurred when getting exit code of process %d\n", m_pid);
+      CloseHandle(processHandle);
       return false;
     }
+    CloseHandle(processHandle);
     if (exitCode != 9999)
     {
       ret = static_cast<int>(exitCode);
